+++ packages/api/src/Domain/Response/SharedVaultUsers/DesignateSurvivorResponse.ts
@@ -1,0 +1,3 @@
+export type DesignateSurvivorResponse = {
+  success: boolean
+}