+++ packages/features/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.59.20](https://github.com/standardnotes/app/compare/@standardnotes/features@1.59.19...@standardnotes/features@1.59.20) (2023-09-26)
+
+**Note:** Version bump only for package @standardnotes/features
+
 ## [1.59.19](https://github.com/standardnotes/app/compare/@standardnotes/features@1.59.18...@standardnotes/features@1.59.19) (2023-09-20)
 
 **Note:** Version bump only for package @standardnotes/features