+++ packages/responses/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.13.44](https://github.com/standardnotes/app/compare/@standardnotes/responses@1.13.43...@standardnotes/responses@1.13.44) (2023-09-26)
+
+**Note:** Version bump only for package @standardnotes/responses
+
 ## [1.13.43](https://github.com/standardnotes/app/compare/@standardnotes/responses@1.13.42...@standardnotes/responses@1.13.43) (2023-09-20)
 
 **Note:** Version bump only for package @standardnotes/responses